backend/requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.10
Flask-Session
neo4j
psycopg2-binary>=2.9
python-dotenv>=1.0.0
APScheduler